+++ src/main/kotlin/no/nav/familie/ba/sak/statistikk/internstatistikk/InternStatistikkController.kt
@@ -1,6 +1,9 @@
 package no.nav.familie.ba.sak.statistikk.internstatistikk
 
+import no.nav.familie.ba.sak.common.RessursUtils
+import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingSøknadsinfoService
 import no.nav.familie.ba.sak.kjerne.behandling.domene.BehandlingÅrsak
+import no.nav.familie.ba.sak.kjerne.behandling.domene.SøknadsstatistikkForPeriode
 import no.nav.familie.ba.sak.sikkerhet.SikkerhetContext
 import no.nav.familie.kontrakter.felles.Ressurs
 import no.nav.security.token.support.core.api.ProtectedWithClaims
@@ -11,17 +14,21 @@ import org.springframework.http.ResponseEntity
 import org.springframework.validation.annotation.Validated
 import org.springframework.web.bind.annotation.GetMapping
 import org.springframework.web.bind.annotation.RequestMapping
+import org.springframework.web.bind.annotation.RequestParam
 import org.springframework.web.bind.annotation.RestController
+import java.time.LocalDate
+import java.time.format.DateTimeParseException
 
 @RestController
-@RequestMapping("/api")
+@RequestMapping("/api/internstatistikk")
 @ProtectedWithClaims(issuer = "azuread")
 @Validated
 class InternStatistikkController(
     private val internStatistikkService: InternStatistikkService,
-) {
+    private val behandlingSøknadsinfoService: BehandlingSøknadsinfoService,
+    ) {
 
-    @GetMapping(path = ["internstatistikk"], produces = [MediaType.APPLICATION_JSON_VALUE])
+    @GetMapping(produces = [MediaType.APPLICATION_JSON_VALUE])
     fun hentAntallFagsakerOpprettet(): ResponseEntity<Ressurs<InternStatistikkResponse>> {
         logger.info("${SikkerhetContext.hentSaksbehandlerNavn()} henter internstatistikk")
         val internstatistikk = InternStatistikkResponse(
@@ -31,6 +38,21 @@ class InternStatistikkController(
             antallBehandlingerPerÅrsak = internStatistikkService.finnAntallBehandlingerPerÅrsak(),
         )
         return ResponseEntity.ok(Ressurs.Companion.success(internstatistikk))
+    }
+
+    @GetMapping(path = ["antallSoknader"], produces = [MediaType.APPLICATION_JSON_VALUE])
+    fun hentSøknadsstatistikkForPeriode(
+        @RequestParam fom: String?,
+        @RequestParam tom: String?,
+    ): ResponseEntity<Ressurs<SøknadsstatistikkForPeriode>> {
+        return try {
+            val fomDato = fom?.let { LocalDate.parse(it) } ?: LocalDate.now().minusMonths(4).withDayOfMonth(1)
+            val tomDato = tom?.let { LocalDate.parse(it) } ?: fomDato.plusMonths(4).minusDays(1)
+
+            RessursUtils.ok(behandlingSøknadsinfoService.hentSøknadsstatistikk(fomDato, tomDato))
+        } catch (e: DateTimeParseException) {
+            RessursUtils.badRequest("Ugyldig dato", e)
+        }
     }
 
     companion object {
